requirements.txt
pygame==1.9.3
cx_Freeze==4.3.4
pyyaml==3.12
